+++ MapDriveInator/scripts/MapDriveInator.groovy
@@ -48,15 +48,16 @@ if(baseFolderNode){
     if(!nodeNewImports) return 'Interrupted: no MDI styles'
     def visibilizarAvance = MDI.wantToLog(nodeNewImports)
     if(visibilizarAvance) texto.append("\n").append('(elapsed time in miliseconds)').append("\n").append((tIni - new Date().getTime()) as String).append("\n")
-    baseFolderNode.style.name = 'baseFolder'
+    baseFolderNode.style.name = MDI.styleBaseFolder
+    baseFolderNode.noteContentType = 'markdown'
     baseFolderPath = MDI.getPathFromLink(baseFolderNode)
     deleteNodesWithLinkToOther(nodeNewImports)
     if(visibilizarAvance) texto.append((tIni - new Date().getTime()) as String).append("\n")
         
     // brought this to before getting the changes done 
-    def linkType = MDI.getLinkType(baseFolderNode)
     def nameFilt = MDI.getFilter(baseFolderNode)
     def maxD = MDI.getMaxDepth(baseFolderNode)
+    def linkType = MDI.getLinkType(baseFolderNode)
     def markMovedOption = MDI.markWhenMoved(baseFolderNode)
     def checkIfBroken = MDI.checkIfReallyBroken(baseFolderNode)
 
@@ -232,7 +233,7 @@ if(baseFolderNode){
     }
     xClonesPend4 = xClonesPend.clone()
 
-    def nodosConFileEnOtraParte = (nodosSinFileB + nodosSinFileC).findAll{new File(it.link).exists()}  //TODO: 
+    def nodosConFileEnOtraParte = (nodosSinFileB + nodosSinFileC).findAll{new File(it.link).exists()}
     // ui.informationMessage(nodosConFileEnOtraParte as String)
     def nodosSinFileBC = (nodosSinFileB + nodosSinFileC) - nodosConFileEnOtraParte
     // ui.informationMessage(nodosSinFileBC as String)
@@ -293,9 +294,9 @@ if(baseFolderNode){
         // agregar f a nodo nueva importación
         def nodoDonde = nodeNewImports
         def gPath = baseFolderPath
-        (f - baseFolderPath)?.split(Pattern.quote(File.separator)).init().each{String dir ->   //TODO: linux
+        (f - baseFolderPath)?.split(Pattern.quote(File.separator)).init().each{String dir ->   
             // ui.informationMessage('dir   :' + dir as String)
-            gPath +=  dir << File.separator  //TODO: linux
+            gPath +=  dir << File.separator
             // ui.informationMessage('gPath   :' + gPath as String)
             nodoDonde = nodoDonde.children.find{it.text == dir}?:nodoDonde.createChild()
             if(nodoDonde.text==''){
@@ -526,7 +527,7 @@ if(baseFolderNode){
 
     //region: ---------------------- Reporte Y Final Main ------------------------------
 
-    if(modoDebug) ui.informationMessage('---------------------- Reporte Y Final Main ------------------------------');
+    if(modoDebug) ui.informationMessage('---------------------- Reporte Y Final Main ------------------------------')
     textoReport.append((((new Date().getTime() - tIni)/100).toInteger()/10) as String).append(" seconds\n\n")
     ui.informationMessage(textoReport.toString())
     textoReport << '=====================================\n\n' << texto
@@ -534,11 +535,11 @@ if(baseFolderNode){
     texto.setLength(0)
     textoReport.setLength(0)
     if (newFilesImported) c.select(nodeNewImports);
-    c.statusInfo = '    -------------   Map-Drive-Inated    -------------      ';
+    MDI.statusInfo('    -------------   Map-Drive-Inated    -------------      ')
     // c.select(baseFolderNode);
 //end:  ------------------------------------------------------------------
 }else{
-    ui.informationMessage("couldn't find the current 'baseFolderNode' or assign a new one \n\n (path between the selected node and the map's root)")
+    return
 }
 
 //end: ============================== fin: MAIN SCRIPT ============================================================
@@ -558,7 +559,7 @@ def armaListadoRutas(nodo, String path){
         if(MDI.nodeIsFolder(it)){
             MDI.markAsMoved(it,false)
             def pathF = MDI.getFolderpathFromStrings(path,it)
-            xFolders << new xFile(it.id, MDI.getPathFromLink3(it,File.separator), pathF) //TODO: linux
+            xFolders << new xFile(it.id, MDI.getPathFromLink3(it,File.separator), pathF)
             if(it.children.size()>0){
                 armaListadoRutas(it, pathF)
             }